+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Pa
 - Gradle-Files überarbeiten (Deprecated-Aufrufe raus)
 - auf Gradle 9 und Kotlin 2.3 aktualisieren
 
+[Unreleased]
+### Added
+- Text um limitTo(..)-Methode ergänzt
+
 ## [6.6.2] -2026-04-05
 ### Fixed
 - weitere polnische Sonderzeichen für Text-Konvertierung nach ASCII berücksichtigt